reldgram: remove all locks but one
This modification simplifies the design of the library that now uses a single lock for everything.
The cost of this is to introduce unnecessary waits that can reduce the performance (e.g., every server
connection shares the same lock).

This must be improved in the future, but at this point I prefer simplicity.